problem-scraper/pyproject.toml
[project]
name = "problem-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
pythonpath = "src"
